[primer-app] Use recommended syntax for React function components.
The TypeScript + React cheatsheet recommends a) adding a `JSX.Element`
type annotation to React function component declarations; and b)
declaring them as `const` values.

This commit changes our components to match this recommended
style. (Note that the components we imported from Tailwind UI needed
the most work, therefore we'll need to be cognizant of this when
importing future componenents from that library.)

See:
https://github.com/typescript-cheatsheets/react#function-components