[primer-app] Switch to React 17-style JSX.
See:

https://reactjs.org/blog/2020/09/22/introducing-the-new-jsx-transform.html

For the moment, the best way to support this in Vite (`esbuild`,
really) is via this configuration:

https://github.com/vitejs/vite/issues/2369#issuecomment-904689947

so that's what we use.

There's a Vite plugin that addresses this here:
https://github.com/alloc/vite-react-jsx. However, that plugin has a
performance impact on production builds because has to use Babel,

The simpler `jsxInject` workaround described here:

https://github.com/vitejs/vite/issues/2369

is not quite as foolproof as the workaround we're using.