[primer-app] Support absolute path aliases for module imports.
Up until now, we've had to use filesystem-relative paths when
importing modules. This is both awkward and fragile — relative imports
may break when source code gets moved around in the filesystem.

With this change, we can now import using paths like "@/..." and
"@components/...". We use both Vite's and TypeScript's support for
path aliases.

Note that, unfortunately, both the Vite and TypeScript path aliasing
systems need to be maintained separately, because Vite doesn't use
TypeScript aliases, and vice versa. Mercifully, for eslint there's a
plugin that uses TypeScript's settings, and we've enabled that plugin
here; so while maintaining 2 configurations is painful, at least it's
not three.