Add primer-components package to the workspace.
The idea is that we can develop these components separately from the
app, so that:

* We can use things like Storybook.js to develop the components
without making the app dependent on Storybook.js, which tends to move
at a slower pace than the rest of the React ecosystem.

* Designers can help design our components using HTML, CSS (Tailwind),
and a bit of React, without needing to know how the app works/is
structured.

This commit just clones all the components in primer-app and then
exports them via an ES Modules module, so effectively this is just
getting me learning how to build a library using Vite & TypeScript.
This isn't yet hooked up to the app, which still has its own copies of
these components.