[primer-components] Storybook.js support for our components.
This uses the storybook-builder-vite Storybook.js builder plugin,
which is a bit raw, but I did manage to get it working. Expect some
rockiness for the next few months, however.

(The alternative is to use Storybook.js's standard builder, which is
Webpack. However, that would require that we replicate all the Vite
resolvers, plugins, etc. with Webpack equivalents, which isn't
feasible.)