Fix up Nix flake derivations for the new packaging.
This was tricky to get right. Between Nix, yarn2nix, TypeScript,
eslint, and esbuild/Vite, and all the various plugins we're using,
there are many moving parts. The trickiest bit was making the module
resolver work for TypeScript type checking in
`hackworthltd-primer-app`. Be very careful before changing any
`tsconfig.json` settings in any of our packages!

I could not get the `hackworthltd-primer-lib` package to work at all,
so I removed it. It's no great loss, as we were only using it to
common-up some config file settings for eslint and TypeScript.

Note that in the Nix derivation for the
`hackworthltd-primer-components` package, we take a shortcut to reduce
the amount of time it takes to create it, by taking advantage of the
fact that the package doesn't have any exported dependencies. All of
its dependencies are either `peerDependencies`, which the user of the
package (in our case, the `hackworthltd-primer-app` package) is
responsible for installing themselves, or `devDependencies`, which are
not necessary at all for downstream dependencies. For some
reason (possibly related to creating a Nix shell?), yarn2nix bundles
`devDependencies`, and this leads to a very large and very
time-consuming closure. Therefore, we override `installPhase` and skip
the `node_modules` bundling for this package.

Likewise, for the `hackworthltd-primer-app` package, we only care
about the `dist` directory created by the `yarn build` step, as this
is what we'll distribute on the CDN/whatever. Therefore, we override
the `installPhase` and `distPhase` for this package, as well.

Finally, note that due to the fact that we're using the
`vite-plugin-checker` plugin for Vite, `yarn build` will run a full
TypeScript and eslint pass on the project when generating the dist
files. Therefore, we don't need a separate set of `-checks` packages,
and I've removed those here.