[primer-components] Add Storybook.js Nix package.
Note that in order to make this work, I had to promote the packages
`peerDependencies` ("react" and "react-dom") to `dependencies`. This
is due to the fact that yarn2nix doesn't know about
`peerDependencies`, and therefore doesn't make them available at
package build time in the Nix expression.

This doesn't appear to cause any problems in either our Yarn build
system, nor with the Nix packages, but if we ever do publish these
packages, we should probably revisit this, as technically, we could
cause problems for downstream consumers of the
"@hackworthltd/primer-components package"".