[meta] Proof-of-concept Chromatic deployment workflow.
Eventually this should run only when triggered by a successful Hydra
build of the project, but for now, we just do it on demand.

I'd like to have the workflow get the Chromatic project token from our
Vault servers, and the EC2 runners have the correct Vault permissions
now to do so, but we'd need to supply it as a regular environment
variable, which doesn't get automatically obscured in workflow log
output like GitHub does with secrets, so there's a bit of a risk in
doing this. Therefore, for now, we give the project token to GitHub
and take advantage of the GitHub Actions secrets infrastructure.