[meta] Test the Chromatic workflow.
This commit makes an innocuous change to one of the React components
in order to test that Chromatic's visual diff CI tests are
working. (They are now.)

This commit also includes more fixes for the Chromatic GitHub Action,
which took quite a bit of fiddling to get working:

- Our self-hosted EC2 image is based on a relatively bare-bones Amazon
Linux AMI, not a heavily-modified Ubuntu as the official GitHub
runners are. This means that many, many packages are missing from the
default install, and at least some GitHub Actions require one or more
of these missing packages. To make the Chromatic action work, I had to
install `yarn` and `node` into the system path in a workflow step
prior to running the Chromatic action.

- Even after installing these tools, the Chromatic action failed,
complaining that `node_modules` wasn't present. Therefore, after
installing `yarn`, we then run `yarn` from the top-level project
directory. This installs the workspace-wide `node_modules`, which
includes all of the Node.js dependencies that the Chromatic action
needs. Note that these `node_modules` aren't installed from the Nix
cache; they're downloaded and installed by `yarn`. This is fine, even
if it's non-optimal due to the installation taking longer than it
might if we used the Nix cache.

- I couldn't get the `workflow_dispatch` trigger to work correctly, so
I added a `push` trigger, as well. Note that this trigger will ignore
`bors trying` branch pushes, since we don't want to bother with
Chromatic CI tests on that branch. The fact that the
`workflow_dispatch` trigger didn't work may have something to do with
the other changes I made as documented above, as I didn't test it
again after I got the `push` dispatch working.

In the long run, we'd prefer to use the `workflow_dispatch` trigger,
combined with a GitHub webhook handler, so that we can wait until
Hydra is done building the Storybook package before we attempt to
deploy to Chromatic. As things stand, the `push` trigger will run as
soon as there's a push to any branch, meaning that if any Nix packages
need to be rebuilt, there's a good chance the EC2 runner will run more
or less simultaneously with the Hydra build, which is obviously not
optimal.

Anyway, this preferred workflow will take some more work, the least of
which is setting up a GitHub webhook service to trigger GitHub
workflows upon Hydra CI events.