style(components): Insert a smaller size into the UIButton hierarchy.
The new "sm" size is now smaller than any previous button size, we add
a new "2xl" size that is the same size as the previous "xl", and all
the other `Size` values get bumped down the hierarchy one size
smaller.

This is done so that UIButton is consistent with Tailwind CSS's
responsive size names.