feat(components): Add an avatar component for the student.
Based on:

https://github.com/hackworthltd/tailwind-ui/blob/4081cc176c7e48a5aed94de35e1d1b46c02f2ac9/react/components/application-ui/elements/avatars/rounded_avatars.jsx

If the student has provided an image, we use that. Otherwise, we use
one of several generated avatars (the student can select the generated
style).

Note that we'll have to be mindful of the "identicon" and (especially)
the "jdenticon" generators, to ensure they don't inadvertently
generate an image that could be perceived as inappropriate or
offensive. I don't believe either generator would do that on purpose,
of course, but as they're basically RNGs, we should be cautious.