style(components): Switch to circular avatars.
This is in keeping with recent web design trends. I originally went
with a square icon because I didn't think it would be easy to generate
round avatars using @dicebear/avatars, but it turns out they added
this capability recently:

https://github.com/dicebear/dicebear/issues/43#issuecomment-570961813

Therefore the new avatar component is based on this Tailwind UI
component:

https://github.com/hackworthltd/tailwind-ui/blob/4081cc176c7e48a5aed94de35e1d1b46c02f2ac9/react/components/application-ui/elements/avatars/circular_avatars.jsx

which is just an `<img>` element.