ci: Add a "forcing function" package for Hydra builds.
We recently switched our Hydra back to the upstream version and
dropped our patches. One consequence of this is that, because upstream
Hydra doesn't report to GitHub the status of cached builds, Bors
wasn't triggering merges if the "required" Hydra job was cached.

To work around this, we add a new dummy package that simply echoes to
stdout the project's git rev. Because this package is included in the
required job, it's guaranteed to trigger a new Hydra build for this
job on Bors merge attempts, since Bors creates a new merge commit for
these attempts. This should fix our issue.