feat(components): Change "user" to "account" in sessions navbar.
Note that we can't use "student" here because both students and
instructors will use this part of the application, and I don't want
this to be misleading if we say "student". "Account" seems like a good
replacement for the more general term, "user," at least so far as
account settings and the like are concerned.