feat(components): Add a notifications bell icon.
This components was factored out of the sessions nav bar component.
Note that in the sessions nav bar, it was part of an anchor (`<a>`)
element in mobile mode, but here we implement only the `<button>`
version.

It's not yet clear to me whether this should be a separate
button/icon, or a decoration on the account avatar.