ci: Chromatic workflow needs to run on push.
I couldn't get this to work on the correct PR using the
workflow_dispatch trigger. The workflow would run and deploy correctly
to Chromatic, but Chromatic didn't seem to know which PR to report the
status back to.

This does mean we'll need to build the Storybook using `nix-build`,
and we might not benefit from the S3 Nix cache, at least not in every
case.

I've also undone some of the hacks I needed to get this working on our
EC2 runners, since we're not using those anymore. (I had missed these
in the previous commits where I disabled our EC2 runners.)