TreeVisx: add oversimplistic node labels
SVG puts later elements on top of earlier ones, so to get labels to show
up properly they need to appear above links (as they overlap).

We do not currently size the node background to fit around the text, as
it is not trivial to do (it is possible, see getComputedTextLength).

We also do not care about label lengths (and thus node size) when
computing the layout.