Remove Visual Studio Code workspace configuration
This reverts 11b0b0064e1e2d8ef187bb867663d744fd7d0428.

It seems that this config isn't very amenable to being checked in to version control, seeing as each user is likely to end up with local modifications. For example, the `folders` field will be modified by adding a folder to the workspace.