fix: Fix the Nix Storybook build.
For some reason, using a tsconfig reference to "../primer-types"
breaks the Nix Storybook build, because the storybook build looks for
`../primer-types` during the Nix build, but it's not there, as
`yarn2nix` appears to ignore the reference. If we remove the
reference, both the interactive and Nix Storybook builds work fine.

What's odd is that if we remove the similar reference blocks from the
`@hackworthltd/primer-app` tsconfig, then the interactive primer-app
build breaks when primer-components tries to import its internal
imports using the special `@/...` import syntax. So it seems that the
tsconfig reference blocks are not necessary unless we use that special
import syntax.

Therefore, we leave them in place for primer-app but remove them for
primer-components.