style(`SessionsPage`): Scroll the session list, pin the navbar.
This was quite tricky to get right!

Note that in our app, `SessionsPage` works as described above (I
hacked things up a bit to test it, but those changes are not ready to
be committed yet). However, in Storybook, it's possible to slightly
scroll the page, and hence the navbar, even though the navbar should
be fixed in place and not scrollable. There's some sort undesirable
interaction happening between our markup and Storybook's.

It might be possible to fix this in Storybook via some custom config,
but for now, keep in mind that `SessionsPage` won't render precisely
the same in Storybook as it will in the actual app.

(Note also that on Safari, it's possible to "bounce" the sessions page
a bit, in an elastic kind of way, whereas on Firefox it's rock solid
and locked into place. This appears to be a common complaint on
StackOverflow, and is in fact present in our Vonnegut prototype's
sessions list page, which otherwise exhibits exactly the behavior we
want. We might be able to do something about Safari's behavior via
some JavaScript hacks, if it's important to fix.)