feat: Add `wsrun` for cross-package watched builds.
This utility can execute a Yarn command on a given package and its
dependencies, and unlike `ultra`, you can configure it to wait for a
particular bit of output on each project before it launches the
command on the next project. This means that we can use `wsrun` to
launch "watch" builds on a given package: `wsrun` will launch
simultaneous, separate Vite "watch" builds for the package and its
dependencies.

This appears to work really well for `@hackworthltd/primer-app`,
because Vite watches the object files of its dependencies. It does not
appear to work very well (if at all) for the Storybook in
`@hackworthltd/primer-components`, but this commit does add
functionality for that anyway.

See `README.md` for more details. I also added some notes to the
README about how `wsrun`, `ultra`, and `tsc --build` differ in terms
of how well they work with our project.