TreeInteractive: better, fixed, layout
To stop things moving around when visualising trees using the
Tree-Interactive/Both component, we display the two renders
side-by-side, with the svg first as it is a fixed size.

We also change the outline view to use put bullet points inside the
bounding box of the <ul> so doesn't overlap the svg.