feat: Add a placeholder "no match" page.
Note that this page will return a 200 OK, because it appears that in
React Router v6, there's no way to specify a particular code for a
given route. That's probably OK, as only browsers will be interacting
with the frontend app, anyway, and they don't report the status code
to the user.