Add D3 to interactive trees
Note that the D3 tree is not itself interactive. It is useful to be able
to see the same tree in multiple renderings, even if they are not all
fully-featured.

We also add a margin to the svg and change the size of the story
renders, to more closely match the visx styling.