feat: start animating D3 trees
This is only half an implementation, and has some obvious bad behaviour.
We animate existing nodes and edges changing position, but immediately
insert new nodes and edges. This leads to momentarily disconnected
trees: the new edge's start poisition is its source nodes' new position,
but the source node is being animated moving from its old position into
this new position.

One solution would be to only insert new nodes and edges after the
transition is finished. However, in some follow-up commits we will take
a different tack: let's animate the insertion of new nodes and edges.
We'll insert them "collapsed" at their parent's (old/pre-animation)
position, and in one animation we will move the existing parent, and
"grow" the new child. This requires a bit of plumbing work, as we need
to remember the old position of the parent.

We also leave any animation of removal of subtrees to future work.