fix: D3 trees display static stories better
Before this commit, when we initially display a non-trivial tree, it
looked rather odd: all the nodes would appear in the correct place, and
then edges would animate "growing from source to target", all in
parallel. This was because we had no cached oldPos, so getPos would
default to the newly computed position.

We change the behaviour of getPos to default to the computed position of
the root. Thus on initial render, the tree seems to "zoom in" from the
root. By this I mean that initially you see one node, at the location of
the root. Next you will see a small version of the tree based at the
root (short edges, but normal-size nodes, as we do not scale anything).
This will then expand to a full-sized tree, with all the nodes and edges
moving in parallel.