feat: Generate Primer API bindings using Orval.
In addition to generating OpenAPI bindings for the Primer API, Orval
will also automatically generate React Query bindings for us, which
saves a lot of boilerplate that we'd otherwise have to hand-write in
order to adapt the openapi-generator Axios bindings to React Query's
API.

(React Query support in the app will come in a subsequent step.)

Note that for now we commit the generated files and also the Primer
OpenAPI JSON spec that we used to generate them. Eventually we might
not want to check them in, but this is simpler, for now.