feat: Add a new `SimplePaginationBar` component.
Adapted from:
https://github.com/hackworthltd/tailwind-ui/commit/4081cc176c7e48a5aed94de35e1d1b46c02f2ac9#diff-4999e0cb5e44af9684a334903aa162d6ffa091875bb7dd9b2e15cf3c2f104a7a

I don't think the UX for this component is great: for example, there's
no way to jump to the beginning or end of the list of results.
However, this component is straightforward to implement, and we have
more pressing needs at the moment.

Note that the `onClick` handlers are included in the props, but aren't
wired up to anything yet, as we need to add support for them to
`UIButton` first :D