feat: Add `onClick` handlers to anything with a `UIButton` component.
Note: because we've enabled the (recommended)
`exactOptionalPropertyTypes` setting in our `tsconfig.ts`
project-wide, it doesn't seem to be generally possible to use an
option type for our optional `onClick` handlers as they propagate down
the component hierarchy. This is pretty annoying and perhaps we should
disable this setting. I'm not quite sure what it's buying us.

ref:
https://www.typescriptlang.org/tsconfig#exactOptionalPropertyTypes