feat(sessions): Hide previous and next page buttons as needed.
On the first and last pages of the list of sessions, hide the previous
and next page buttons, respectively.

This needs some styling work. I think that the ideal scenario would be
to leave a blank button-sized placeholder `div` so that the "Previous"
and "Next" buttons are always in the same place, regardless of whether
the other button is visible. However, it's not clear how to do that
and still remove the button from the accessibility tree.