fix: Hide the sessions page summary if there are no sessions.
This change addresses a bug, where previously the summary would read
"Showing 1 to 0 of 0 sessions" if there were no sessions. Our fix here
is simply to hide the summary line altogether in this case. This
behavior is consistent with our plans to display some kind of
special-purpose call to action when the student has no sessions.