feat: Add a grouped button list component
This will form the basis of a "what do you want to insert" screen, where
we may want to list constructors grouped by what type they are from, or
values grouped by type. We don't implement the "how to group" logic
here (indeed, it is not clear exactly what this logic should be).