Re-generate the Primer API.
This looks like a lot of changes, but it is mainly down to the
upstream JSON API spec having be re-ordered. (This is pretty annoying,
and we should probably investigate whether it's possible to make the
order more deterministic.)

There does appear to be at least one semantic change: the `term`
property in the `Def` interface is now an optional. However, this does
not impact the primer-app code base as we don't use it yet.

Note: I've also disabled the Orval-generated mock API tests, as we
don't have the infrastructure in place to use them, and they throw up
a bunch of errors. We can easily re-generate these in the future if we
decide we want them.