fix: When deploying to Chromatic, just run it interactively.
Prior to this change, we built a copy of Storybook using Nix, then
passed that static Storybook directory to Chromatic for publishing.
However, as of Storybook 6.4.22 (at least), the `build-storybook`
functionality of Storybook.js doesn't work anymore in our Nix build,
failing with cryptic errors on an operation that used to work fine.

We don't use the Nix Storybook derivation for anything other than
publishing to Chromatic in CI. In interactive use, we always build
Storybook in our Nix shell, live and with HMR. Therefore, there's not
really any point in building the Nix derivation containing the
Storybook, so this change just builds the Storybook as a one-off when
we push to Chromatic in CI.

One ramification of this change is that if there are builds that don't
modify the Storybook output, then they'll always rebuild the Storybook
anyway rather than reusing the previous result from the Nix cache.
However, in practice building the Storybook is reasonably fast (for
now, at least), so this is not a huge price to pay.