feat: Customize styling of React Flow nodes and edges
This starts bringing us closer to our designs, but it's still a work in progress. For one thing, we don't have colours, seeing as we don't yet have access to the appropriate metadata for distinguishing different types of nodes.

The `nodeWidth` and `nodeHeight` parameters now actually set the visible dimensions of nodes, rather than being used solely for layout.

`extra-styles.css` is a hack around an upstream issue: . Fortunately we'll be able to remove it once we switch to using a custom node.