feat: Allow multiple React Flow trees to be drawn on one canvas
We remove `onClick` handling for now. We no longer have a single `onClick` per canvas, so it would be arbitrary to choose just one. And things are made difficult by the fact that `primer-types` doesn't currently attach a definition name to trees. We're waiting on an update to `primer-types` to make click handlers be per-node anyway.