fix: Disable Chromatic CI checks temporarily.
We do, however, add a new job that builds the project's Storybook, to
ensure that's working.

This change is due to https://github.com/hackworthltd/primer-app/pull/339,
or, more specifically, this comment:

https://github.com/hackworthltd/primer-app/pull/339#issuecomment-1142255802

We've already spent more time than we can probably justify trying to
fix this, so it's time to punt. Ideally we'd like to file an issue
upstream, but this may not be possible to reproduce in a way that we
can share with them.

Note that we're not losing *Storybook* functionality with this commit,
only the Chromatic review process. The local Storybook still works
fine in `pnpm watch:storybook` mode.