fix: Component-ise some functions in order to fix React Hooks issue
The key here is that calling a function does not create a subcomponent, as far as React is concerned. The semantics are the same as inlining the function. So we have to use JSX syntax in order to make the contents of the T&D tab actual components (as a prerequisite, we must refactor these functions to take a single object argument, rather than multiple arguments). This means that React doesn't get confused about the number of hooks in use varying when we switch between tabs.

It's a little surprising that this wasn't picked up by the React Hooks ESLint plugin.