refactor: Inline `TabContents`
This makes it clear that it is not a React component, whereas other top-level functions returning `JSX.Element` in this module are.

The syntax is slightly odd because JS doesn't allow `if` or `switch` in _expressions_ (though there are proposals that would solve this: https://github.com/tc39/proposal-do-expressions, https://github.com/tc39/proposal-pattern-matching). The standard solution is to apply a no-argument function, as we have done here (unless the scrutinee is a boolean, in which case we can use the ternary operator, as we do elsewhere in the module).