feat: Proof-of-concept that we can pull in `primer` Nix packages.
Mainly we want to demonstrate in this commit that we can pull `primer`
Nix packages into the `primer-app` flake without needing to rebuild
them from scratch. The reason this is important to demonstrate is that
`primer` packages take a long time to build, especially if we need to
build, say, GHC, and therefore we want `primer-app` to take full
advantage of our `primer` Nix binary cache.

And in fact, in my testing, building the
`packages.x86_64-linux.primer-service-docker-image` package does pull
the package from the Nix binary cache, rather than rebuilding it.