chore: Switch Storybook config to TypeScript.
Note that there's some breakage here wrt. the vite-plugin-svgr plugin,
but we can work around it with a `require`-style import. We also need
to tell `ts-node` to use CommonJS when generating modules, or else
`build-storybook` will complain about using imports outside of
modules. (Sigh.) In any case, this should have no effect on generated
code for the app, because it doesn't use `ts-node`.