chore: Upgrade to react-query v4.
I used the migration notes for this upgrade, but nearly everything was
handled automatically by Orval, as it integrates with React Query.

Notes:

* The package is now called `@tanstack/react-query`. Orval will only
work with v4 if you use the new name, so that's what we've done here:

https://github.com/anymaniax/orval/issues/497#issuecomment-1192445575

* v4 encourages users to migrate to the new signal-based query
cancellation mechanism as described here:

https://tanstack.com/query/v4/docs/guides/query-cancellation

We've implemented that here, and in the process we've enabled Orval's
support for this feature, removing the old `cancelToken` approach in
the process. (This has not been well-tested, but then neither was the
old way.)