fix: Update to work with new `Tree` type
We commit to ReactFlow and drop our other tree renderers, rather than updating them to work with the new API.

`packages/primer-types/src/Tree.ts` is a mess of copy-pasted type definitions from `primer-app`. But soon this will no longer be necessary as we'll fix our project structure / build system.

This is enough to get the codebase to typecheck again, but is still a work-in-progress:
- patterns not rendered
- directions ignored i.e. right child treated as just another child - this means `match` expressions extend too far downward
- definition name and type not shown
- various color issues - in particular, many nodes are black