refactor: Remove `useLayout` options
This makes the code cleaner and more type-safe.

There's no point staying close to the implementation from the ReactFlow example for the sake of it, now that we've already diverged quite far, and will soon go further. Note that using a different direction produced visually undesirable results anyway, seeing as other parts of the codebase implicitly assumed top-bottom edges.