feat: Use `patternsUnder` style for trees
This was actually implemented in the backend precisely because Dagre doesn't give us any way to align children horizontally.

We also update the "odd and even" example to reflect this. But we keep the old trees around. As far as the dumb frontend is concerned, they're still valid, so we may as well ensure that we continue to render them sensibly. And while we've moved on from the rendering of patterns without application nodes, the other difference, patterns as right-children, can still be used by passing different flags to the backend.