chore(nodejs): Bump dependencies.
Note that this bump includes a new version of `@heroicons/react`,
which changes the appearance of the outline icons.

Also, many of the icon names have changed. Here's a convenient list of
the changes:

https://github.com/tailwindlabs/heroicons/issues/750

Finally, the imports now need to specify the icon size.

We do not bump the TypeScript pin here because `@dicebear/avatars` has
some issues with strict null checks in TypeScript 4.8. See:

https://github.com/dicebear/dicebear/issues/239