feat: Add click handler and selection to `TreeReactFlow`
The initial intention here was that the component would use an internal hook to keep track of selection.

But seeing as our app will also allow selection changes to be effected from elsewhere (e.g. the definitions list), we would still need a selection parameter so that we could set it externally. And at that point, tracking the state within the component becomes pointless. We would have two state hooks at different layers tracking the same thing.