refactor: Move everything into a single package.
We do this for the following reasons:

* We want to get rid of the `primer-types` package, which duplicates
types exported from the Primer API. However, in the process of doing
this, we ran into some long-standing upstream issues (described in
https://github.com/hackworthltd/primer-app/pull/545) which seem to be
related to using React Hooks defined in one package but used in
another.

* I originally chose this particular project organization so that we
could decouple Storybook.js from the frontend application, because I
was worried that, if they were integrated into the same package,
issues with Storybook and new versions of React would hold up
development of the frontend. However, we've already successfully made
the transition from React v17 to v18, and Storybook.js was ready
before we were, so this hasn't been an issue in practice.

* Interactive development would be simpler, and possibly faster, if we
unified the packages.

The refactoring here could probably be improved, but this was the
simplest first step I could come up with. The next most obvious step
is to remove the `primer-app` workspace and move to just a flat
package.

Note that I have *not* actually removed the `primer-types` types in
this commit: I've only moved them into a single `Types.ts` file.