fix: Add a project-local `.envrc` file.
This fixes the issue described here:

https://github.com/hackworthltd/primer-app/pull/547#issuecomment-1273060657

It does so by configuring `direnv` to write its cache for this project
in `~/.cache/direnv/layouts`. This is a bit intrusive, but without
this workaround, `vite` commands search the local `.direnv` cache for
dependencies, and it adds about 10 seconds or more to every `vite`
command. Note that the fix comes from here:

https://github.com/direnv/direnv/wiki/Customizing-cache-location

Before falling back to this workaround, I also tried the following
Vite configuration tricks, but none of them worked:

https://vitejs.dev/config/server-options.html#server-watch
https://vitejs.dev/config/dep-optimization-options.html#optimizedeps-exclude

There are multiple (somewhat) related issues currently filed on Vite
which suggest that convincing Vite not to look everywhere in the
project root directory is a known issue:

https://github.com/vitejs/vite/pull/8778/files
https://github.com/vitejs/vite/issues/7363
https://github.com/vitejs/vite/issues/8341